requirements.txt
Flask==3.0.3
google-api-python-client==2.142.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
gunicorn==21.2.0
